requirements.txt
Flask
gunicorn
pytesseract
PyPDF2
Pillow
PyCryptodome
spacy
pytextrank